# [WIP] Optimize evaluation of float arguments (version 2)