--- a/sympy/core/mod.py
+++ b/sympy/core/mod.py
@@ -37,6 +37,7 @@ class Mod(Function):
     """
 
     kind = NumberKind
+    _skip_eval_float_arguments = False
 
     @classmethod
     def eval(cls, p, q):