--- a/sympy/functions/elementary/integers.py
+++ b/sympy/functions/elementary/integers.py
@@ -129,6 +129,7 @@ class floor(RoundFunction):
 
     """
     _dir = -1
+    _skip_eval_float_arguments = False
 
     @classmethod
     def _eval_number(cls, arg):