<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs

See the discussion at #23565.

Benchmark master:

```
from sympy import cos, atan
from random import random

%timeit atan(random())
542 µs ± 9.09 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)


%timeit cos(random())
393 µs ± 4.35 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```
Benchmark PR
```
%timeit atan(random())
89.2 µs ± 1.5 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

%timeit cos(random())
90.4 µs ± 1.8 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```


#### Brief description of what is fixed or changed

Optimize evaluation of float arguments 

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* functions
   * Optimize evaluation of float arguments. A new class attribute `_skip_eval_float_arguments` is added to the `Function` class. If set to `True`, then on numeric evaluation of the class the `eval` is skipped.
<!-- END RELEASE NOTES -->