@oscarbenjamin 2022-06-06
I really want to see some proper benchmarks for things like this. The benchmark suite needs to be improved. It currently doesn't handle caching properly so I don't know if it can meaningfully measure the effect of something like this.