@asmeurer 2022-06-06
See my comment here https://github.com/sympy/sympy/pull/23565#issuecomment-1147978591. I am currently -1 on this because:

- I don't see the need for a new flag. `_should_evalf` should suffice. Also `_should_evalf` currently defaults to evalfing on Float but it could be redefined to evalf anything, so naming the flag in this way is misleading. 
- Calling `_should_evalf` (and `evalf`) before `eval` is a behavior change that could affect custom functions.
- This seems to optimize the case of floating-point arguments over the case of symbolic arguments. But symbolic arguments are far more common, so they should be the case that is more optimized. 
- The current logic doesn't require any code to redefine `_should_evalf`. With this, the flag has to be added to several functions. This suggests that the current behavior is a better default, and furthermore, that making a change like this would likely break any user code that is similar.