@eendebakpt 2022-06-07
@asmeurer I agree with your observation that sympy should optimize for symbolic arguments and I share your concerns about changing behavior.

To give a bit of context: this PR was started because there are some methods with an `eval` that is really expensive. In #23541 I added a fast path for floats and @oscarbenjamin suggested an alternative optimization that is generic and not (as in # 23541) specific for one particular function. That suggestion was worked out to this PR.

* The changing behaviour is the main concern here. For this reason the attribute `_skip_eval_for_float` was added so we can disable the new behaviour for functions that rely on `eval` (in this PR this was required for about 10 functions to make all tests pass). Or, we can set `_skip_eval_for_float=False` by default and only enable it for functions we have checked to be safe.

* The check `_should_evalf` is really fast compared to the `eval`. So we can gain a lot on the evaluation of floats, while having more or less the same performance on symbolic expressions (in particular because the function evaluations are cached). There are already some benchmarks in #23565 and some open PRs to improve performance independent of this PR. 
I would be happy to perform more benchmarking, but the main point is the change of behaviour.

@asmeurer @oscarbenjamin Based on your comments I gather that having `_skip_eval_for_float=True` is not an option due to the change in behaviour. 
We can still add the mechanism with `_skip_eval_for_float` and allow this for specific functions. This would allow #23541 to remove the special `imaginary_unit_as_coefficient` method. It that an option? If not, I will close the PR and continue with #23541.