@oscarbenjamin 2022-06-07
> I gather that having `_skip_eval_for_float=True` is not an option due to the change in behaviour.

We have to consider the fact that downstream code somewhere that subclasses Function probably depends on the current behaviour and we should try to preserve compatibility for that.